// src/interfaces/wms/company_wms.ts
export interface ICompany {
  company_code: string;
  company_name: string;
  address1: string;
  address2: string;
  address3: string;
  city: string;
  country: string;
  bl_section_1: string;
  bl_section_2: string;
  stmt_of_accounts: string;
  app_license_001: string;
  picking_path: string;
  mail_server: string;
  mail_email: string;
  mail_pwd: string;
  bill_auth_pwd: string;
}